Correction after review

+++ main.sh
@@ -5,10 +5,18 @@
 
 set -e
 
-source .lib/utilis.sh
-
 ./scripts/update.sh
-./config_lang.sh
-./script/config_user.sh
+.scripts/config_lang.sh
+./scripts/config_user.sh
 
-sudo reboot
+until [ "$reboot" = "n" ]
+do
+    echo "Reboot? y/n"
+    read reboot
+    if [ "$reboot" = 'y' ]
+    then
+        echo "System will restart"
+        sudo reboot
+    fi
+done
+echo "Back to Home"
